# web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "enum": [
              "invalid_input",
              "too_long",
              "quota_exceeded",
              "rate_limited",
              "unauthorized",
              "busy",
              "not_found",
              "internal"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "JobCreate": {
        "properties": {
          "blend_ratio": {
            "default": 0.5,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Blend Ratio",
            "type": "number"
          },
          "blend_voice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blend Voice"
          },
          "lang": {
            "default": "a",
            "description": "Language code, see /v1/voices",
            "title": "Lang",
            "type": "string"
          },
          "pitch": {
            "default": 0.0,
            "maximum": 6.0,
            "minimum": -6.0,
            "title": "Pitch",
            "type": "number"
          },
          "pronunciations": {
            "items": {
              "$ref": "#/components/schemas/Pronunciation"
            },
            "maxItems": 200,
            "title": "Pronunciations",
            "type": "array"
          },
          "speed": {
            "default": 1.0,
            "maximum": 2.0,
            "minimum": 0.5,
            "title": "Speed",
            "type": "number"
          },
          "text": {
            "description": "Text to speak (limit depends on the account)",
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "turnstile_token": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required for anonymous visitors",
            "title": "Turnstile Token"
          },
          "voice": {
            "default": "af_heart",
            "title": "Voice",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "JobCreate",
        "type": "object"
      },
      "JobOut": {
        "properties": {
          "audio_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Seconds"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "estimated_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Seconds"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "progress": {
            "default": 0,
            "title": "Progress",
            "type": "integer"
          },
          "queue_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Position"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "done",
              "error",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "stream_url": {
            "title": "Stream Url",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Signed MP3 link, valid for an hour (when done)",
            "title": "Url"
          },
          "wav_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Signed WAV link (signed-in users)",
            "title": "Wav Url"
          }
        },
        "required": [
          "id",
          "status",
          "stream_url"
        ],
        "title": "JobOut",
        "type": "object"
      },
...
      "Pronunciation": {
        "properties": {
          "say": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Say",
            "type": "string"
          },
          "word": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Word",
            "type": "string"
          }
        },
        "required": [
          "word",
          "say"
        ],
        "title": "Pronunciation",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Text-to-speech jobs with live streaming. Errors always look like `{\"error\": {\"code\": ..., \"message\": ...}}`.",
    "title": "Narravo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/tts/jobs": {
      "post": {
        "description": "Queue text for speech. Follow it live on `stream_url`, or poll the job.\n\nAnonymous visitors must include a Cloudflare Turnstile token. Signed-in users'\nsaved pronunciations are applied unless the request brings its own.",
        "operationId": "create_job_v1_tts_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Job",
        "tags": [
          "tts"
        ]
      }
    },
    "/v1/tts/jobs/{job_id}": {
      "delete": {
        "description": "Ask the worker to stop. The final status arrives as a `cancelled` event.",
        "operationId": "cancel_job_v1_tts_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "tts"
        ]
      },
      "get": {
        "description": "Status, progress and (when done) a fresh download link.",
        "operationId": "get_job_v1_tts_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Job",
        "tags": [
          "tts"
        ]
      }
    },
...
  }
}